Demo env and readme

+++ src/services/backup.js
@@ -5,6 +5,7 @@ import EmailSender from '@services/emailSender';
 import Club from '@models/club';
 import User from '@models/user';
 import Role from '@models/role';
+import Suggestion from '@models/suggestion';
 
 import ApiError from '@utilities/apiError';
 import errorCodes from '@config/errorCodes';
@@ -15,11 +16,13 @@ class Backup {
       const users = await User.find({});
       const clubs = await Club.find({});
       const roles = await Role.find({});
+      const suggestions = await Suggestion.find({});
 
       const finalJson = {
         clubs,
         users,
         roles,
+        suggestions,
       };
 
       const fileName = `${moment().format('YYYY-MM-DD HH-mm-ss')}.json`;
@@ -42,20 +45,36 @@ class Backup {
     }
   };
 
-  restore = async (fileName) => {
+  restore = async (fileName, collectionsToRetore = []) => {
     try {
       const fileData = await fs.promises.readFile(`backups/${fileName}`, 'utf-8');
 
-      const { users, clubs, roles } = JSON.parse(fileData);
+      const {
+        users,
+        clubs,
+        roles,
+        suggestions,
+      } = JSON.parse(fileData);
 
-      await Role.deleteMany({});
-      await Role.insertMany(roles);
+      if (collectionsToRetore.includes('roles')) {
+        await Role.deleteMany({});
+        await Role.insertMany(roles);
+      }
+      
+      if (collectionsToRetore.includes('users')) {
+        await User.deleteMany({});
+        await User.insertMany(users);
+      }
 
-      await User.deleteMany({});
-      await User.insertMany(users);
+      if (collectionsToRetore.includes('clubs')) {
+        await Club.deleteMany({});
+        await Club.insertMany(clubs);
+      }
 
-      await Club.deleteMany({});
-      await Club.insertMany(clubs);
+      if (collectionsToRetore.includes('suggestions')) {
+        await Suggestion.deleteMany({});
+        await Suggestion.insertMany(suggestions);
+      }
     } catch (error) {
       throw new ApiError(errorCodes.Internal, error);
     }
